Implemented swipe to delete on projects
